git: Extract `worker::git` module
